+++ RELEASES.md
@@ -1,3 +1,42 @@
+# Release 0.37.2 (2024-04-16)
+
+- Do not include legend box if there is no legend ([#58](https://github.com/Axect/Peroxide/pull/58)) (Thanks to [@GComitini](https://github.com/GComitini))
+- Add `rtol` field to `BroydenMethod`
+- Implement high-level macros for root finding
+  - `bisection!(f, (a,b), max_iter, tol)`
+  - `newton!(f, x0, max_iter, tol)` (require `#[ad_function]` attribute)
+  - `secant!(f, (a,b), max_iter, tol)`
+  - `false_position!(f, (a,b), max_iter, tol)`
+
+# Release 0.37.1 (2024-04-15)
+
+- Implement `BrodenMethod`: Broyden's method (`I>=1, O>=1, T=([f64; I], [f64; I])`)
+- Restore citation file
+
+# Release 0.37.0 (2024-04-14)
+
+## Huge Update - Whole new Root finding & anyhow
+
+### Whole new Root finding
+
+- Remove all boilerplates
+- Now, `RootFinding` is composed of traits
+  - `RootFindingProblem<const I: usize, const O: usize, T>`: Trait for defining and root finding problem
+    - `I`: Input dimension
+    - `O`: Output dimension
+    - `T`: Type of state
+  - `RootFinder`: Trait for finding root
+    - `BisectionMethod`: Bisection Method (`I=1, O=1, T=(f64, f64)`)
+    - `FalsePositionMethod`: False Position Method (`I=1, O=1, T=(f64, f64)`)
+    - `NewtonMethod`: Newton Method (`I=1, O=1, T=f64`)
+    - `SecantMethod`: Secant Method (`I=1, O=1, T=(f64, f64)`)
+
+### Error handling with anyhow
+
+- Remove `thiserror` dependency
+- Add `anyhow` for error handling
+- Change error handling in `ODE`, `Spline`, `WeightedUniform`
+
 # Release 0.36.4 (2024-04-11)
 
 - More generic Butcher tableau